Generate real mouse events on mouseMove
Actual mouseMove events are necessary to properly simulate the
behavior of complex components such as Flickable. Instead of just
manipulating the cursor position, record the current button state so
that we can generate move events that are not suppressed.

In addition, add a mouseStartDrag() convenience function so that client
code does not need to be aware of QML's interpretation of the drag
threshold value.

Change-Id: I1ded27f976846660a0ec91906d0b2a6372efccd5
Reviewed-by: Robin Burchell <robin+qt@viroteck.net>